remove Battlefield bot opponent && the automatically arrangement of ships transfer to the constructor_field
