+++ README.md
@@ -62,7 +62,6 @@ coming soon...]: #
 Планы на будущее:
 ------
 - Заменить BattlefieldBotOpponent на BattlefieldOpponent. Автоматическую генерацию выделить в отдельный класс
-- добавить варианты установки для macos (вообще придумать что-то более мультиплатформенное)
 - обновить uml-диаграмму проекта
 - научиться открывать поля в разных размерах и на полный экран (скорее всего не выйдет из-за особенностей tkinter)
 - переписать сервер на flask
