+++ application_game_with_bot.py
@@ -3,8 +3,8 @@ from tkinter import messagebox
 
 import config
 from battlefield_bot import BattlefieldBotPlayer
-from battlefield_bot_view import BattlefieldBotOpponent
 from battlefield_player_view import BattlefieldPlayer
+from battlefield_opponent_view import BattlefieldOpponent
 from constructor_fields import ConstructorFields
 from wrappers import Window
 from wrappers import Canvas
@@ -27,30 +27,26 @@ class Application:
         self.__start_loops()
 
     def __create_board(self, is_need_for_randomness):
-        constructor_field = ConstructorFields(presence_timer=False, is_need_for_randomness=is_need_for_randomness)
+        constructor_field = ConstructorFields(is_need_for_randomness=is_need_for_randomness)
         constructor_field = constructor_field.get()
         if not constructor_field:
             self.__is_closing = True
             return
         self.__window = Window(is_game_field=True)
-        self.__canvas = Canvas(self.__window,small_canvas=False)
+        self.__canvas = Canvas(self.__window, number_of_fields=2)
         # the order in which the fields are created is very important
         # it has to do with filling the canvas with buttons
-        self.__foreign_field = BattlefieldBotOpponent(is_only_generation=False, canvas=self.__canvas)
-        self.__field = BattlefieldPlayer(constructor_field + [], self.__canvas)
-
-        self.__bot_field = BattlefieldBotPlayer(constructor_field + [])
+        auto_field = ConstructorFields(is_need_for_randomness=True)
+        self.__foreign_field = BattlefieldOpponent(real_field=auto_field.get(), canvas=self.__canvas)
+        self.__field = BattlefieldPlayer(constructor_field, self.__canvas)
+        self.__bot_field = BattlefieldBotPlayer(constructor_field)
 
     def __draw_boards(self, my_first_move):
         if self.__is_closing:
             return
-        self.__field.view()
-        self.__foreign_field.view()
-        constant_number_of_frame_at_the_bottom = 3  # 1 for each field and 1 for turn
+
         label_frame = tk.Frame(self.__canvas,
-                               width=(2 * config.COLUMN + 1) * config.SIZE_OF_CELL - 2 * (
-                                       (2 * config.COLUMN + 1) * config.SIZE_OF_CELL
-                                       // constant_number_of_frame_at_the_bottom),
+                               width=(2 * config.COLUMN + 1) * config.SIZE_OF_CELL // 3,
                                height=config.SIZE_OF_CELL,
                                bg="white")
         label_frame.pack_propagate(False)
@@ -68,12 +64,14 @@ class Application:
             self.__bot_field.take_a_shot()
         self.__label_turn.pack(fill="both", expand=True)
         self.__canvas.create_window(
-            ((2 * config.COLUMN + 1) * config.SIZE_OF_CELL // constant_number_of_frame_at_the_bottom,
+            ((2 * config.COLUMN + 1) * config.SIZE_OF_CELL // 3,
              config.ROW * config.SIZE_OF_CELL),
             anchor="nw",
             window=label_frame)
-        self.__canvas.pack()
+        self.__foreign_field.view()
+        self.__field.view()
         self.__window.resizable(width=False, height=False)
+        self.__canvas.pack(fill="both", expand=True)
 
     def __start_loops(self):
         if self.__is_closing:
@@ -88,11 +86,11 @@ class Application:
         self.__check_game_over(n)
 
     def __click_processing(self):
-        if self.__foreign_field.presence_of_changes() and self.__label_turn["fg"] == "green2":
+        if self.__foreign_field.get_changes() and self.__label_turn["fg"] == "green2":
             if not self.__foreign_field.existence_hit_last_shot():
                 self.__label_turn.configure(text="Ход противника", fg="red")
                 self.__bot_field.take_a_shot()
-        elif self.__bot_field.presence_of_changes() and self.__label_turn["fg"] == "red":
+        elif self.__bot_field.get_changes() and self.__label_turn["fg"] == "red":
             # objects in the tkinter can update more slowly than this cycle, so:
             self.__window.update()
             self.__canvas.after(config.TIME_WAITING_BOT_MOVE)
